scripts/06_pipeline_health.py
#!/usr/bin/env python3
"""
Script 6: Pipeline Health Monitor
====================================

Analyzes all active deals in Attio, checks for stalls and gaps, calculates
pipeline metrics, and generates alerts to Slack + Google Drive.

Checks per deal:
  - Days in current stage vs. expected velocity
  - Missing fields (ai_account_brief, contacts, deal amount)
  - Last activity date (email, meeting, note)
  - Stall signals based on stage-specific thresholds

Metrics:
  - Total pipeline value by stage
  - Pipeline coverage ratio
  - Average deal velocity by stage
  - Win rate trending

Stage Velocity Benchmarks:
  Lead: 14 days max, 7-day stall alert
  Intro Call: 7 days max, 5-day stall alert
  Discovery: 21 days max, 10-day stall alert
  Solutioning: 30 days max, 14-day stall alert
  Redlines: 21 days max, 7-day stall alert

Outputs:
  - Slack summary with actionable alerts
  - Detailed Google Drive report

Triggers:
  - Weekly Monday mornings
  - On-demand
"""

import argparse
import logging
import yaml
from datetime import datetime
from pathlib import Path

CONFIG_DIR = Path(__file__).parent.parent / "config"
TEMPLATE_DIR = Path(__file__).parent.parent / "templates"
logger = logging.getLogger(__name__)


class PipelineHealthMonitor:
    """Monitors pipeline health and generates alerts."""

    def __init__(self, attio_client, slack_client, gdrive_client):
        self.attio = attio_client
        self.slack = slack_client
        self.gdrive = gdrive_client
        self.stage_config = self._load_stage_config()

    def _load_stage_config(self):
        with open(CONFIG_DIR / "pipeline_stages.yaml") as f:
            return yaml.safe_load(f)

    def get_active_deals(self):
        """Query all active deals in Attio (stages 1-5, 8)."""
        pass

    def check_deal_health(self, deal):
        """
        Analyze a single deal for health issues:
        - Days in stage vs. benchmark
        - Missing required fields
        - Activity recency
        - Stall signals
        Returns list of alerts/issues.
        """
        pass

    def calculate_pipeline_metrics(self, deals):
        """
        Calculate aggregate metrics:
        - Total value by stage
        - Coverage ratio
        - Average velocity
        - Win rate trend
        """
        pass

    def generate_alerts(self, deal_health_results):
        """
        Generate actionable alert messages:
        - "Deal X in Discovery 23 days, no activity — schedule follow-up?"
        - "Deal Y missing technical contact — run Buying Committee Builder?"
        - "3 deals in Redlines with no close date"
        """
        pass

    def post_to_slack(self, summary, alerts):
        """Post pipeline health summary + alerts to Slack."""
        pass

    def save_to_gdrive(self, full_report):
        """Save detailed pipeline health report to Google Drive."""
        pass

    def run(self):
        """Execute full pipeline health check."""
        deals = self.get_active_deals()
        health_results = [self.check_deal_health(d) for d in deals]
        metrics = self.calculate_pipeline_metrics(deals)
        alerts = self.generate_alerts(health_results)
        self.post_to_slack(metrics, alerts)
        self.save_to_gdrive({"metrics": metrics, "alerts": alerts, "details": health_results})


def main():
    parser = argparse.ArgumentParser(description="Pipeline Health Monitor")
    parser.add_argument("--output", choices=["slack", "gdrive", "both", "console"],
                        default="both")
    parser.add_argument("--dry-run", action="store_true")
    args = parser.parse_args()

    logging.basicConfig(level=logging.INFO)

    monitor = PipelineHealthMonitor(
        attio_client=None, slack_client=None, gdrive_client=None
    )
    monitor.run()


if __name__ == "__main__":
    main()
