requirements.txt
# Onboarded GTM Engine Dependencies
pyyaml>=6.0
requests>=2.28
python-dotenv>=1.0
anthropic>=0.40
google-auth>=2.0
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
slack-sdk>=3.0
